[SPARK-44017][PS][CONNECT][TESTS] Reorganize `ps.Groupby` unit tests
### What changes were proposed in this pull request?
Reorganize `ps.Groupby` related unit tests, break them into several topics

### Why are the changes needed?
to make them easy to maintain

### Does this PR introduce _any_ user-facing change?
No, test-only

### How was this patch tested?
updated CI

Closes #41541 from zhengruifeng/ps_tests_reorg_group.

Authored-by: Ruifeng Zheng <ruifengz@apache.org>
