+++ dev/sparktestsupport/modules.py
@@ -753,8 +753,16 @@ pyspark_pandas_slow = Module(
         "pyspark.pandas.tests.frame.test_time_series",
         "pyspark.pandas.tests.frame.test_truncate",
         "pyspark.pandas.tests.io.test_io",
-        "pyspark.pandas.tests.test_groupby",
-        "pyspark.pandas.tests.test_groupby_slow",
+        "pyspark.pandas.tests.groupby.test_aggregate",
+        "pyspark.pandas.tests.groupby.test_apply_func",
+        "pyspark.pandas.tests.groupby.test_cumulative",
+        "pyspark.pandas.tests.groupby.test_describe",
+        "pyspark.pandas.tests.groupby.test_groupby",
+        "pyspark.pandas.tests.groupby.test_head_tail",
+        "pyspark.pandas.tests.groupby.test_index",
+        "pyspark.pandas.tests.groupby.test_missing_data",
+        "pyspark.pandas.tests.groupby.test_split_apply",
+        "pyspark.pandas.tests.groupby.test_stat",
         "pyspark.pandas.tests.test_indexing",
         "pyspark.pandas.tests.test_ops_on_diff_frames",
         "pyspark.pandas.tests.test_ops_on_diff_frames_slow",
@@ -954,8 +962,16 @@ pyspark_pandas_slow_connect = Module(
         "pyspark.pandas.tests.connect.frame.test_parity_time_series",
         "pyspark.pandas.tests.connect.frame.test_parity_truncate",
         "pyspark.pandas.tests.connect.io.test_parity_io",
-        "pyspark.pandas.tests.connect.test_parity_groupby",
-        "pyspark.pandas.tests.connect.test_parity_groupby_slow",
+        "pyspark.pandas.tests.connect.groupby.test_parity_aggregate",
+        "pyspark.pandas.tests.connect.groupby.test_parity_apply_func",
+        "pyspark.pandas.tests.connect.groupby.test_parity_cumulative",
+        "pyspark.pandas.tests.connect.groupby.test_parity_describe",
+        "pyspark.pandas.tests.connect.groupby.test_parity_groupby",
+        "pyspark.pandas.tests.connect.groupby.test_parity_head_tail",
+        "pyspark.pandas.tests.connect.groupby.test_parity_index",
+        "pyspark.pandas.tests.connect.groupby.test_parity_missing_data",
+        "pyspark.pandas.tests.connect.groupby.test_parity_split_apply",
+        "pyspark.pandas.tests.connect.groupby.test_parity_stat",
         "pyspark.pandas.tests.connect.test_parity_indexing",
         "pyspark.pandas.tests.connect.test_parity_ops_on_diff_frames",
         "pyspark.pandas.tests.connect.test_parity_ops_on_diff_frames_slow",
